+++ src/reducers/index.ts
@@ -1,28 +1,7 @@
 import { combineReducers } from 'redux';
 
-import { UPDATE_COUNT } from '~/constant';
-import { CounterActionTypes } from '~/actionTypes';
-import { CountState } from '~/stateTypes';
+import * as reducers from './reducers';
 
-import { account } from '~/reducers/account';
-
-export const initialCounter: CountState = { count: 0 };
-
-export const counter = (
-    state = initialCounter,
-    action: CounterActionTypes
-): CountState => {
-    switch (action.type) {
-        case UPDATE_COUNT:
-            return { ...state, count: action.payload };
-        default:
-            return state;
-    }
-};
-
-export const rootReducer = combineReducers({
-    counter,
-    account
-});
+export const rootReducer = combineReducers(reducers);
 
 export type RootState = ReturnType<typeof rootReducer>;
