+++ src/reducers/reducers.ts
@@ -1,0 +1,2 @@
+export { account } from '~/reducers/account';
+export { counter } from '~/reducers/counter';
